+++ scripts/bkenv.properties
@@ -283,6 +283,8 @@ BK_CI_OPENAPI_API_PORT=21935
 BK_CI_PLUGIN_API_PORT=21925
 # BK_CI_PROCESS_API_PORT默认为21921. 无需修改.
 BK_CI_PROCESS_API_PORT=21921
+# BK_CI_ENGINE_API_PORT目前开源没有独立部署，先占坑默认为21917. 无需修改.
+BK_CI_ENGINE_API_PORT=21917
 # BK_CI_PROJECT_API_PORT默认为21912. 无需修改.
 BK_CI_PROJECT_API_PORT=21912
 # BK_CI_QUALITY_API_PORT默认为21928. 无需修改.
