--- src/backend/ci/core/audit/api-audit/src/main/kotlin/com/tencent/devops/audit/api/pojo/Audit.kt
@@ -1,53 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making BK-CI 蓝鲸持续集成平台 available.
- *
- * Copyright (C) 2019 THL A29 Limited, a Tencent company.  All rights reserved.
- *
- * BK-CI 蓝鲸持续集成平台 is licensed under the MIT license.
- *
- * A copy of the MIT License is included in this file.
- *
- *
- * Terms of the MIT License:
- * ---------------------------------------------------
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
- * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
- * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of
- * the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT
- * LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN
- * NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
- * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package com.tencent.devops.audit.api.pojo
-
-import io.swagger.annotations.ApiModel
-import io.swagger.annotations.ApiModelProperty
-
-@ApiModel("审计模型-Audit")
-data class Audit(
-    @ApiModelProperty("资源类型", required = true)
-    val resourceType: String,
-    @ApiModelProperty("资源ID", required = true)
-    val resourceId: String,
-    @ApiModelProperty("资源名称", required = true)
-    val resourceName: String,
-    @ApiModelProperty("操作人", required = true)
-    val userId: String,
-    @ApiModelProperty("操作", required = true)
-    val action: String,
-    @ApiModelProperty("操作内容", required = true)
-    val actionContent: String,
-    @ApiModelProperty("项目id", required = true)
-    val projectId: String
-) {
-    companion object {
-        const val classType = "audit"
-    }
-}
