+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/pojo/audit/AuditPage.kt
@@ -25,11 +25,26 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-subprojects {
-    dependencies {
-        compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
-        compile "org.jetbrains.kotlin:kotlin-reflect"
-        testCompile "junit:junit"
-        testCompile "org.springframework.boot:spring-boot-starter-test"
-    }
+package com.tencent.devops.process.pojo.audit
+
+import io.swagger.annotations.ApiModel
+import io.swagger.annotations.ApiModelProperty
+
+@ApiModel("Audit分页数据包装模型")
+data class AuditPage<out T>(
+    @ApiModelProperty("总记录行数", required = true)
+    val count: Long,
+    @ApiModelProperty("第几页", required = true)
+    val page: Int,
+    @ApiModelProperty("每页多少条", required = true)
+    val pageSize: Int,
+    @ApiModelProperty("总共多少页", required = true)
+    val totalPages: Int,
+    @ApiModelProperty("数据", required = true)
+    val records: List<T>,
+    @ApiModelProperty("是否拥有创建权限", required = true)
+    val hasCreatePermission: Boolean
+) {
+    constructor(page: Int, pageSize: Int, count: Long, records: List<T>, hasCreatePermission: Boolean) :
+        this(count, page, pageSize, Math.ceil(count * 1.0 / pageSize).toInt(), records, hasCreatePermission)
 }
