+++ src/backend/ci/settings.gradle
@@ -228,9 +228,3 @@ include ":core:sign:biz-sign"
 include ":core:sign:biz-sign-sample"
 include ":core:sign:boot-sign"
 include ":core:sign:model-sign"
-
-include ":core:audit"
-include ":core:audit:api-audit"
-include ":core:audit:biz-audit"
-include ":core:audit:boot-audit"
-include ":core:audit:model-audit"
