+++ support-files/sql/0001_ci_create-database_mysql.sql
@@ -16,4 +16,3 @@ CREATE DATABASE IF NOT EXISTS `devops_ci_op` DEFAULT CHARACTER SET utf8mb4;
 CREATE DATABASE IF NOT EXISTS `devops_ci_quality` DEFAULT CHARACTER SET utf8mb4;
 CREATE DATABASE IF NOT EXISTS `devops_ci_openapi` DEFAULT CHARACTER SET utf8mb4;
 CREATE DATABASE IF NOT EXISTS `devops_ci_sign` DEFAULT CHARACTER SET utf8mb4;
-CREATE DATABASE IF NOT EXISTS `devops_ci_audit` DEFAULT CHARACTER SET utf8mb4;
