--- support-files/sql/1001_ci_audit_ddl_mysql.sql
@@ -1,24 +1,0 @@
-USE devops_ci_audit;
-
-SET NAMES utf8mb4;
-SET FOREIGN_KEY_CHECKS = 0;
-
--- ----------------------------
--- Table structure for T_AUDIT_RESOURCE
--- ----------------------------
-
-CREATE TABLE IF NOT EXISTS `T_AUDIT_RESOURCE` (
-  `ID` bigint(20) NOT NULL AUTO_INCREMENT,
-  `RESOURCE_TYPE` varchar(32) NOT NULL,
-  `RESOURCE_ID` varchar(128) NOT NULL,
-  `RESOURCE_NAME` varchar(128) NOT NULL,
-  `USER_ID` varchar(64) NOT NULL DEFAULT '',
-  `ACTION` varchar(64) NOT NULL DEFAULT '',
-  `ACTION_CONTENT` varchar(1024) NOT NULL DEFAULT '',
-  `CREATED_TIME` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-  `UPDATED_TIME` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-  `IS_DELETED` bit(1) DEFAULT b'0',
-  `STATUS` varchar(32) DEFAULT NULL,
-  `PROJECT_ID` varchar(128) NOT NULL,
-  PRIMARY KEY (`ID`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4
