--- support-files/templates/#etc#ci#application-audit.yml
@@ -1,10 +1,0 @@
-# Image Service Template 服务配置文件模板
-spring:
-  datasource:
-    url: jdbc:mysql://__BK_CI_MYSQL_ADDR__/devops_ci_audit?useSSL=false&autoReconnect=true&timezone=+800&useUnicode=true&characterEncoding=utf8&allowMultiQueries=true&sessionVariables=sql_mode=%27STRICT_TRANS_TABLES,NO_ZERO_IN_DATE,NO_ZERO_DATE,ERROR_FOR_DIVISION_BY_ZERO,NO_AUTO_CREATE_USER,NO_ENGINE_SUBSTITUTION%27
-    username: __BK_CI_MYSQL_USER__
-    password: __BK_CI_MYSQL_PASSWORD__
-
-# 服务器端口配置，在同一台机器上部署多个微服务，端口号要不同 21936
-server:
-  port: __BK_CI_AUDIT_API_PORT__
